refactor(financial-components): remove unused financial chart components to enhance maintainability and reduce code complexity
