+++ resources/js/components/FinancialChart.vue
