+++ resources/js/components/FinancialDashboard.vue
