+++ resources/js/components/dashboard/FinancialSummaryDoughnut.vue
