+++ resources/js/components/dashboard/KeyTrends.vue
